+++ config/routes.rb
@@ -8,6 +8,9 @@ Rails.application.routes.draw do
 
     api_version(module: 'V2', path: { value: 'v2' }) do
       constraints(id: UUID_REGEX) do
+        namespace :admin do
+          resources :impersonation, only: :create
+        end
         resources :account_lists, only: [:index, :show, :update] do
           scope module: :account_lists do
             resource :analytics, only: [:show]